// src/api/models/YogsScheduleSchema.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Jingle Jam Community Page Public API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Stream } from './Stream';
import {
    StreamFromJSON,
    StreamFromJSONTyped,
    StreamToJSON,
} from './Stream';
import type { YogsScheduleSchemaDaysInner } from './YogsScheduleSchemaDaysInner';
import {
    YogsScheduleSchemaDaysInnerFromJSON,
    YogsScheduleSchemaDaysInnerFromJSONTyped,
    YogsScheduleSchemaDaysInnerToJSON,
} from './YogsScheduleSchemaDaysInner';

/**
 * 
 * @export
 * @interface YogsScheduleSchema
 */
export interface YogsScheduleSchema {
    /**
     * 
     * @type {string}
     * @memberof YogsScheduleSchema
     */
    title: string;
    /**
     * 
     * @type {Date}
     * @memberof YogsScheduleSchema
     */
    start: Date;
    /**
     * 
     * @type {Date}
     * @memberof YogsScheduleSchema
     */
    end: Date;
    /**
     * 
     * @type {number}
     * @memberof YogsScheduleSchema
     */
    initialDayIndex: number;
    /**
     * 
     * @type {Array<YogsScheduleSchemaDaysInner>}
     * @memberof YogsScheduleSchema
     */
    days: Array<YogsScheduleSchemaDaysInner>;
    /**
     * 
     * @type {Array<Stream>}
     * @memberof YogsScheduleSchema
     */
    streams: Array<Stream>;
}

/**
 * Check if a given object implements the YogsScheduleSchema interface.
 */
export function instanceOfYogsScheduleSchema(value: object): value is YogsScheduleSchema {
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('start' in value) || value['start'] === undefined) return false;
    if (!('end' in value) || value['end'] === undefined) return false;
    if (!('initialDayIndex' in value) || value['initialDayIndex'] === undefined) return false;
    if (!('days' in value) || value['days'] === undefined) return false;
    if (!('streams' in value) || value['streams'] === undefined) return false;
    return true;
}

export function YogsScheduleSchemaFromJSON(json: any): YogsScheduleSchema {
    return YogsScheduleSchemaFromJSONTyped(json, false);
}

export function YogsScheduleSchemaFromJSONTyped(json: any, ignoreDiscriminator: boolean): YogsScheduleSchema {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'],
        'start': (new Date(json['start'])),
        'end': (new Date(json['end'])),
        'initialDayIndex': json['initialDayIndex'],
        'days': ((json['days'] as Array<any>).map(YogsScheduleSchemaDaysInnerFromJSON)),
        'streams': ((json['streams'] as Array<any>).map(StreamFromJSON)),
    };
}

export function YogsScheduleSchemaToJSON(value?: YogsScheduleSchema | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'title': value['title'],
        'start': ((value['start']).toISOString()),
        'end': ((value['end']).toISOString()),
        'initialDayIndex': value['initialDayIndex'],
        'days': ((value['days'] as Array<any>).map(YogsScheduleSchemaDaysInnerToJSON)),
        'streams': ((value['streams'] as Array<any>).map(StreamToJSON)),
    };
}